[core] Fix thrown error during eval phase
This change changes the way we throw errors for our top-level exported
functions. With the introduction of our build-time container components,
we also started using the top-level APIs in these components. This in
itself is ok since we don't want any restrictions. But this lead to an
issue where an error was thrown, typically if trying to add sx prop on
such container components (1 example).
This was because these container component files were also being
evaluated during the eval phase of wyw and that phase also included
calls to generateAtomics which always throws error.
So with this change, we first detect whether the function is being
called during the eval phase or not, by checking a special global
variable called `__wyw_dynamic_import` which is set by WyW before
evaluation happens. So now these functions don't throw error during the
eval stage.