[plugin] Fix bug related to path separator on Windows
While debugging on Windows, I found that (atleast on Windows 11), the
path separator was `/` instead of the expected `\` resulting in some
files being skipped from being transformed from the `node_modules`.

So the fix first normalizes the path first as per the platform and then
checks for the `transformLibraries`'s presence in the normalized path.