fix: use the correct serializer based on the direction value
Curently, the assumption was that the rtl css would only be needed as an
addon.
Fixed this to always use an explicit direction serializer based on the
direction.
Also, the Vite plugin was not passing on the direction information to
pre-processor. This has been fixed.

Fixes https://github.com/mui/material-ui/issues/44647